requirements.txt
alabaster==0.7.12
appdirs==1.4.4
attrs==19.3.0
Babel==2.8.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
docutils==0.16
flake8==3.8.1
Flask==1.1.2
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.19.0
packaging==20.3
pandas==1.0.5
pathspec==0.8.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-igraph==0.8.2
pytz==2020.1
regex==2020.5.14
requests==2.23.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
texttable==1.6.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
